test/test-requirements.txt
boto3>=1.21.31
pdf2image>=1.16.0
pytesseract>=0.3.9
pytest>=7.1.1
python-docx>=0.8.11
